# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "El Mexicano API",
  "description": "Backend réplica (demo) de **Transportes El Mexicano S.R.L.**: venta de pasajes de bus, rutas e itinerarios, carga y encomiendas, puerta a puerta y centro de ayuda.\n\nFechas en hora de Bolivia (America/La_Paz) y montos en bolivianos (Bs).",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/oficinas": {
   "get": {
    "tags": [
     "Catálogos"
    ],
    "summary": "Boleterías y bodegas con horario",
    "operationId": "oficinas_api_v1_oficinas_get",
    "parameters": [
     {
      "name": "ciudad",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "examples": [
        "Santa Cruz"
       ],
       "title": "Ciudad"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/OficinaOut"
         },
         "title": "Response Oficinas Api V1 Oficinas Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/salidas": {
   "get": {
    "tags": [
     "Salidas y asientos"
    ],
    "summary": "Buscar salidas por origen, destino y fecha",
    "operationId": "buscar_api_v1_salidas_get",
    "parameters": [
     {
      "name": "origen",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Nombre, código o alias de la ciudad",
       "examples": [
        "Sucre"
       ],
       "title": "Origen"
      },
      "description": "Nombre, código o alias de la ciudad"
     },
     {
      "name": "destino",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "examples": [
        "Santa Cruz"
       ],
       "title": "Destino"
      }
     },
     {
      "name": "fecha",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "description": "Por defecto, hoy",
       "title": "Fecha"
      },
      "description": "Por defecto, hoy"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BusquedaSalidasOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/reservas/{codigo}": {
   "get": {
    "tags": [
     "Reservas y pasajes"
    ],
    "summary": "Consultar una reserva",
    "operationId": "consultar_api_v1_reservas__codigo__get",
    "parameters": [
     {
      "name": "codigo",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Codigo"
      }
     },
     {
      "name": "documento",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "description": "Documento del comprador o de un pasajero (protege los datos de la reserva)",
       "title": "Documento"
      },
      "description": "Documento del comprador o de un pasajero (protege los datos de la reserva)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReservaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/puerta-a-puerta": {
   "post": {
    "tags": [
     "Carga y encomiendas"
    ],
    "summary": "Solicitar recojo o entrega a domicilio",
    "description": "Sucre y Santa Cruz, lunes a viernes hábiles. Entregas 08:00–12:00, recojos 14:00–17:00.",
    "operationId": "solicitar_puerta_api_v1_puerta_a_puerta_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SolicitudPuertaIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SolicitudPuertaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/faqs/buscar": {
   "get": {
    "tags": [
     "Centro de ayuda"
    ],
    "summary": "Buscar en las preguntas frecuentes",
    "operationId": "buscar_api_v1_faqs_buscar_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "examples": [
        "puedo llevar a mi perro"
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FaqBusquedaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/bot/encomiendas/{numero_guia}": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "rastrear_encomienda: estado de una guía",
    "description": "Estado, oficina de retiro con dirección y horario, pago pendiente y últimos 3 eventos.\nDatos completos solo si `caller_id` es el teléfono del destinatario o del remitente. Nunca devuelve el PIN.",
    "operationId": "rastrear_encomienda_api_bot_encomiendas__numero_guia__get",
    "parameters": [
     {
      "name": "numero_guia",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Numero Guia"
      }
     },
     {
      "name": "caller_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}",
       "title": "Caller Id"
      },
      "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/encomiendas": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "mis_encomiendas: envíos activos de quien llama",
    "description": "Hasta 5 encomiendas no finalizadas donde `caller_id` es remitente o destinatario.",
    "operationId": "mis_encomiendas_api_bot_encomiendas_get",
    "parameters": [
     {
      "name": "caller_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}",
       "title": "Caller Id"
      },
      "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/salidas": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "consultar_salidas: horarios, precios y asientos libres de un día",
    "operationId": "consultar_salidas_api_bot_salidas_get",
    "parameters": [
     {
      "name": "origen",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Ciudad de origen (acepta alias: 'Santa Cruz de la Sierra', 'SCZ')",
       "title": "Origen"
      },
      "description": "Ciudad de origen (acepta alias: 'Santa Cruz de la Sierra', 'SCZ')"
     },
     {
      "name": "destino",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Ciudad de destino",
       "title": "Destino"
      },
      "description": "Ciudad de destino"
     },
     {
      "name": "fecha",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "hoy, mañana, pasado mañana, un día de la semana, 2026-10-02 o 2/10",
       "title": "Fecha"
      },
      "description": "hoy, mañana, pasado mañana, un día de la semana, 2026-10-02 o 2/10"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/rutas": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "listar_rutas: rutas con distancia, duración y horarios",
    "operationId": "listar_rutas_api_bot_rutas_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/tarifas-carga": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "cotizar_envio: precio de un envío de carga o encomienda",
    "operationId": "cotizar_envio_api_bot_tarifas_carga_get",
    "parameters": [
     {
      "name": "origen",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Origen"
      }
     },
     {
      "name": "destino",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Destino"
      }
     },
     {
      "name": "peso_kg",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Peso en kilos, por ejemplo 3 o 2,5",
       "title": "Peso Kg"
      },
      "description": "Peso en kilos, por ejemplo 3 o 2,5"
     },
     {
      "name": "tipo",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "sobre, paquete o carga; si se omite se deduce del peso",
       "title": "Tipo"
      },
      "description": "sobre, paquete o carga; si se omite se deduce del peso"
     },
     {
      "name": "puerta_a_puerta",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "true si se entrega a domicilio",
       "title": "Puerta A Puerta"
      },
      "description": "true si se entrega a domicilio"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/reservas/{codigo}": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "consultar_reserva: estado de una reserva de pasajes",
    "description": "Estado, boletos, salida y demora. Detalle del viaje y montos solo si `caller_id` es el del comprador.",
    "operationId": "consultar_reserva_api_bot_reservas__codigo__get",
    "parameters": [
     {
      "name": "codigo",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Codigo"
      }
     },
     {
      "name": "caller_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}",
       "title": "Caller Id"
      },
      "description": "Teléfono de quien llama; en ElevenLabs, la variable dinámica {{system__caller_id}}"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/oficinas": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "info_oficinas: dirección, teléfono y horario de las oficinas de una ciudad",
    "operationId": "info_oficinas_api_bot_oficinas_get",
    "parameters": [
     {
      "name": "ciudad",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Ciudad"
      }
     },
     {
      "name": "tipo",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "pasajes o carga (opcional)",
       "title": "Tipo"
      },
      "description": "pasajes o carga (opcional)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/faqs/buscar": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "buscar_faq: respuesta corta a una pregunta frecuente",
    "operationId": "buscar_faq_api_bot_faqs_buscar_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/empresa": {
   "get": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "info_empresa: contactos y dónde comprar pasajes",
    "operationId": "info_empresa_api_bot_empresa_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/bot/puerta-a-puerta": {
   "post": {
    "tags": [
     "Agente de voz (ElevenLabs)"
    ],
    "summary": "solicitar_puerta_a_puerta: agenda un recojo o una entrega a domicilio",
    "description": "Solo Sucre y Santa Cruz, de lunes a viernes hábiles (recojos 14:00–17:00, entregas 08:00–12:00).\nSi falta algún dato responde qué pedirle al cliente. El teléfono de contacto por defecto es `caller_id`.",
    "operationId": "solicitar_puerta_a_puerta_api_bot_puerta_a_puerta_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BotPuertaIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BusquedaSalidasOut": {
    "properties": {
     "ruta": {
      "type": "string",
      "title": "Ruta"
     },
     "origen": {
      "type": "string",
      "title": "Origen"
     },
     "destino": {
      "type": "string",
      "title": "Destino"
     },
     "fecha": {
      "type": "string",
      "format": "date",
      "title": "Fecha"
     },
     "distancia_km": {
      "type": "integer",
      "title": "Distancia Km"
     },
     "duracion_estimada_min": {
      "type": "integer",
      "title": "Duracion Estimada Min"
     },
     "salidas": {
      "items": {
       "$ref": "#/components/schemas/SalidaResumen"
      },
      "type": "array",
      "title": "Salidas"
     },
     "mensaje": {
      "type": "string",
      "title": "Mensaje"
     }
    },
    "type": "object",
    "required": [
     "ruta",
     "origen",
     "destino",
     "fecha",
     "distancia_km",
     "duracion_estimada_min",
     "salidas",
     "mensaje"
    ],
    "title": "BusquedaSalidasOut"
   },
...
   "EstadoVenta": {
    "type": "string",
    "enum": [
     "pendiente_pago",
     "pagada",
     "cancelada",
     "expirada",
     "reembolsada_parcial",
     "reembolsada"
    ],
    "title": "EstadoVenta"
   },
...
   "FaqBusquedaOut": {
    "properties": {
     "consulta": {
      "type": "string",
      "title": "Consulta"
     },
     "resultados": {
      "items": {
       "$ref": "#/components/schemas/FaqOut"
      },
      "type": "array",
      "title": "Resultados"
     },
     "mensaje": {
      "type": "string",
      "title": "Mensaje"
     }
    },
    "type": "object",
    "required": [
     "consulta",
     "resultados",
     "mensaje"
    ],
    "title": "FaqBusquedaOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OficinaOut": {
    "properties": {
     "oficina_id": {
      "type": "integer",
      "title": "Oficina Id"
     },
     "codigo": {
      "type": "string",
      "title": "Codigo"
     },
     "nombre": {
      "type": "string",
      "title": "Nombre"
     },
     "tipo": {
      "$ref": "#/components/schemas/TipoOficina"
     },
     "ciudad": {
      "type": "string",
      "title": "Ciudad"
     },
     "direccion": {
      "type": "string",
      "title": "Direccion"
     },
     "referencia": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Referencia"
     },
     "telefono_e164": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Telefono E164"
     },
     "whatsapp_e164": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Whatsapp E164"
     },
     "url_mapa": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url Mapa"
     },
     "horario_texto": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Horario Texto"
     },
     "telefono": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Telefono",
      "readOnly": true
     }
    },
    "type": "object",
    "required": [
     "oficina_id",
     "codigo",
     "nombre",
     "tipo",
     "ciudad",
     "direccion",
     "referencia",
     "telefono_e164",
     "whatsapp_e164",
     "url_mapa",
     "horario_texto",
     "telefono"
    ],
    "title": "OficinaOut"
   },
...
   "ReservaOut": {
    "properties": {
     "codigo_reserva": {
      "type": "string",
      "title": "Codigo Reserva"
     },
     "estado": {
      "$ref": "#/components/schemas/EstadoVenta"
     },
     "canal": {
      "$ref": "#/components/schemas/CanalVenta"
     },
     "comprador": {
      "type": "string",
      "title": "Comprador"
     },
     "subtotal_bs": {
      "type": "number",
      "title": "Subtotal Bs"
     },
     "descuento_bs": {
      "type": "number",
      "title": "Descuento Bs"
     },
     "total_bs": {
      "type": "number",
      "title": "Total Bs"
     },
     "expira_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Expira At"
     },
     "pagada_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pagada At"
     },
     "salida": {
      "$ref": "#/components/schemas/SalidaDeReserva"
     },
     "boletos": {
      "items": {
       "$ref": "#/components/schemas/BoletoOut"
      },
      "type": "array",
      "title": "Boletos"
     },
     "pago": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PagoOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "factura": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FacturaOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "mensaje": {
      "type": "string",
      "title": "Mensaje"
     }
    },
    "type": "object",
    "required": [
     "codigo_reserva",
     "estado",
     "canal",
     "comprador",
     "subtotal_bs",
     "descuento_bs",
     "total_bs",
     "expira_at",
     "pagada_at",
     "salida",
     "boletos",
     "mensaje"
    ],
    "title": "ReservaOut"
   },
...
   "SolicitudPuertaIn": {
    "properties": {
     "tipo": {
      "$ref": "#/components/schemas/TipoSolicitudPuerta"
     },
     "ciudad": {
      "type": "string",
      "title": "Ciudad",
      "examples": [
       "Sucre"
      ]
     },
     "cliente": {
      "$ref": "#/components/schemas/PersonaIn"
     },
     "telefono": {
      "type": "string",
      "title": "Telefono"
     },
     "direccion": {
      "type": "string",
      "maxLength": 250,
      "minLength": 5,
      "title": "Direccion"
     },
     "referencia": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 250
       },
       {
        "type": "null"
       }
      ],
      "title": "Referencia"
     },
     "latitud": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 90.0,
        "minimum": -90.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitud"
     },
     "longitud": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 180.0,
        "minimum": -180.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitud"
     },
     "fecha_programada": {
      "type": "string",
      "format": "date",
      "title": "Fecha Programada"
     },
     "peso_estimado_kg": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Peso Estimado Kg",
      "description": "Desde 1 kg"
     },
     "descripcion": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 250
       },
       {
        "type": "null"
       }
      ],
      "title": "Descripcion"
     },
     "numero_guia": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Numero Guia",
      "description": "Para entregas: guía de la encomienda"
     }
    },
    "type": "object",
    "required": [
     "tipo",
     "ciudad",
     "cliente",
     "telefono",
     "direccion",
     "fecha_programada"
    ],
    "title": "SolicitudPuertaIn"
   },
...
   "TipoOficina": {
    "type": "string",
    "enum": [
     "boleteria",
     "bodega_carga",
     "mixta"
    ],
    "title": "TipoOficina"
   },
...
   "TipoSolicitudPuerta": {
    "type": "string",
    "enum": [
     "recojo",
     "entrega"
    ],
    "title": "TipoSolicitudPuerta"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    }
   }
  }
 }
}